Revert "fix: update credential list when network is switched"
