+++ packages/react-native/lib/index.tsx
@@ -53,12 +53,9 @@ export const WalletSDKContext = React.createContext<WalletSDKContextProps>({
 });
 
 setStorage(AsyncStorage);
-
 export {useAccounts};
 export {useDIDManagement};
 export {useCredentialUtils, useCredentialStatus};
-export {useDocument, useDocuments} from './documentsHooks';
-
 export function getStorage() {
   return AsyncStorage;
 }
@@ -106,6 +103,75 @@ export function useAccount(address) {
     onDelete,
   };
 }
+
+const useEventListener = (eventManager, eventNames, listener) => {
+  useEffect(() => {
+    eventNames.forEach(eventName => eventManager.on(eventName, listener));
+    return () =>
+      eventNames.forEach(eventName =>
+        eventManager.removeListener(eventName, listener),
+      );
+  }, [eventManager, eventNames, listener]);
+};
+
+const events = [
+  WalletEvents.documentAdded,
+  WalletEvents.documentRemoved,
+  WalletEvents.documentUpdated,
+];
+
+export function useDocument(id) {
+  const [document, setDocument] = useState(null);
+
+  const refetchDocument = useCallback(
+    async updatedDoc => {
+      if (updatedDoc.id !== id) return;
+      const doc = await getWallet().getDocumentById(id);
+      setDocument(doc);
+    },
+    [id],
+  );
+
+  useEffect(() => {
+    getWallet().getDocumentById(id).then(setDocument);
+  }, [id]);
+
+  useEventListener(getWallet().eventManager, events, refetchDocument);
+
+  return document;
+}
+
+export function useDocuments({type}) {
+  const [documents, setDocuments] = useState([]);
+  const [loading, setLoading] = useState(true);
+
+  const fetchDocuments = useCallback(
+    async (updatedDoc, forceFetch = false) => {
+      if (
+        forceFetch ||
+        updatedDoc?.type === type ||
+        updatedDoc?.type?.includes(type)
+      ) {
+        const docs = await getWallet().getDocumentsByType(type);
+        setDocuments(docs);
+        setLoading(false);
+      }
+    },
+    [type],
+  );
+
+  useEffect(() => {
+    fetchDocuments(null, true);
+  }, [fetchDocuments, setLoading]);
+
+  useEventListener(getWallet().eventManager, events, fetchDocuments);
+
+  return {
+    documents,
+    loading,
+  };
+}
+
 export function useWallet() {
   return useContext(WalletSDKContext);
 }
