+++ packages/react-native/package.json
@@ -36,7 +36,7 @@
         "ts-node": "^10.9.1",
         "typescript": "^5.0.4",
         "@testing-library/react-hooks": "^8.0.0",
-        "react-test-renderer": "18.2.0",
+        "react-test-renderer": "17.0.2",
         "react": "^18.2.0",
         "@types/react": "^18.0.24",
         "react-dom": "^17.0.2",
